+++ schema/app/models/user_company_status.py
@@ -1,6 +1,6 @@
 import enum
 
-from sqlalchemy import BigInteger, Column, Enum, Index, String
+from sqlalchemy import Column, Enum, Index, Integer, String
 from sqlalchemy.ext.declarative import declarative_base
 
 Base = declarative_base()
@@ -15,10 +15,10 @@ class Status(enum.Enum):
 class UserCompanyStatus(Base):
     __tablename__ = "user_company_status"
 
-    id = Column(BigInteger, primary_key=True, autoincrement=True)
+    id = Column(Integer, primary_key=True, autoincrement=True)
     user_id = Column(String)
     status = Column(Enum(Status))
-    company_id = Column(BigInteger)
+    company_id = Column(Integer)
 
     __table_args__ = (
         Index("ix_user_company_status_user_id", "user_id"),
